requirements.txt
asyncio==3.4.3
decorator==4.4.2
gevent==1.4.0
greenlet==0.4.15
gssapi==1.6.2
kafka-python==2.0.1
pkg-resources==0.0.0
six==1.14.0
websocket==0.2.1
